added demo button and fixed minor bugs in 40 files

+++ src/Pages/EmployeePages/Employee_Routes.js
@@ -1,7 +1,7 @@
 import { Routes,Route,useNavigate } from "react-router-dom";
-import {useEffect} from "react";
+import {useEffect,useState,useRef} from "react";
 import toast from "react-hot-toast";
-import logo from "../../svgs/Logo.webp";
+import Logo from "../../svgs/Logo.webp";
 import Box from '@mui/material/Box';
 import {loggedInCheckService} from "../LoginPage/loginServices";
 import {NavBar} from "../../BasicComponents/NavBar";
@@ -14,36 +14,36 @@ function EMPLOYEE_ROUTES(){
     
     const navigate = useNavigate();
 
+    const [isAuthorized,setIsAuthorized] = useState(false);
+    const componentMounted = useRef(true);
+
     const Buttons = [
-        {sx:{fontSize:"larger",px:2},
+        {sx:{fontSize:"larger",px:2,borderRadius:"1.5rem"},
         darkColor:"black",
         lightColor:"white",
         heading:"Dashboard",
         value:"dashboard/"},
 
-        {sx:{fontSize:"larger",px:2},
+        {sx:{fontSize:"larger",px:2,borderRadius:"1.5rem"},
         darkColor:"black",
         lightColor:"white",
         heading:"Customers",
         value:"customers/"},
 
-        {sx:{fontSize:"larger",px:2},
+        {sx:{fontSize:"larger",px:2,borderRadius:"1.5rem"},
         darkColor:"black",
         lightColor:"white",
         heading:"My Account",
         value:"my-profile/"},
       ];
 
-    const logos = [logo,{height:"3rem",width:"3rem",borderRadius:"50%",margin:"auto 1%"}];
-    // const [Buttons,setButtons] = useState(defaultButtons);
+    const logo = [Logo,{height:"3rem",width:"3rem",borderRadius:"50%",margin:"auto 1%"}];
+ 
     const ButtonsStyle = {display:"flex",flexDirection:"row",p:2};
-
-    const clicking = (e,value)=>{
-      navigate(value);
-    }
 
     const navStyle = {backgroundColor:"dodgerblue",display:"flex",flexDirection:"row"};
 
+    const navProps = [Buttons,ButtonsStyle,logo,navStyle];
 
     useEffect(()=>{
 
@@ -52,26 +52,25 @@ function EMPLOYEE_ROUTES(){
         const admin_auth = sessionStorage.getItem('admin_auth');
   
         async function loggedInCheck(authorization){
-            const response = await loggedInCheckService(authorization);
-            // const data = await response.json();
-            if(response.status === 200){
-                // toast.success(`you are already logged In as ${data.name} ( ${data.role} )`);
-                // if(data.role === "admin"){
-                //   navigate("/admin/dashboard");
-                // }
-                // else if(data.role === "manager"){
-                //   navigate("/manager/dashboard");
-                // }
-                // else if(data.role === "employee"){
-                //   navigate("/employee/dashboard");
-                // }
+          if(componentMounted.current){
+              const response = await loggedInCheckService(authorization);
+              // const data = await response.json();
+              if(componentMounted.current){
+                   if(response.status === 200){
+                     setIsAuthorized(true);
+                     return;
+                   }else{
+                     // setIsLoading(false);
+                     sessionStorage.clear();
+                     navigate("/login");
+                     return toast.error("couldn't authorize User,please login again");
+                   }
+              }else{
                 return;
-            }else{
-                // setIsLoading(false);
-                sessionStorage.clear();
-                navigate("/login");
-                return toast.error("couldn't authorize User,please login again");
-            }
+              }
+          }else{
+            return;
+          }
         }
           if(employee_auth){
                loggedInCheck({employee_auth});
@@ -86,21 +85,24 @@ function EMPLOYEE_ROUTES(){
     // eslint-disable-next-line
       },[]);
 
-
     return(
+      
         <Box sx={{overflow: "hidden"}}>
-        <Box>
-          <NavBar Buttons={Buttons} ButtonsStyle={ButtonsStyle} clicking={clicking} navStyle={navStyle} logo={logos}/>
+
+          <Box>
+            <NavBar navProps={navProps} isAuthorized={isAuthorized} />
+          </Box>
+
+          {(isAuthorized)
+            ? <Box sx={{overflow:"auto",backgroundColor:"rgba(30, 144, 255, 0.666)",minHeight:"93.5vh",maxHeight:"93.5vh"}}>
+                 <Routes>
+                     <Route path="dashboard/*" element={< EMPLOYEE_DASHBOARD_PAGE />}/>
+                     <Route path="customers/*" element={<EMPLOYEE_CUSTOMER_ROUTES/>}/>
+                     <Route path="my-profile/*" element={<EMPLOYEE_MYPROFILE_PAGE/>}/>
+                 </Routes>
+              </Box>
+            :""}
         </Box>
-        {/* backgroundColor:"dodgerblue" should play with this */}
-        <Box sx={{overflow:"auto",backgroundColor:"rgba(30, 144, 255, 0.666)",minHeight:"93.5vh",maxHeight:"93.5vh"}}>
-          <Routes>
-              <Route path="dashboard/*" element={< EMPLOYEE_DASHBOARD_PAGE />}/>
-              <Route path="customers/*" element={<EMPLOYEE_CUSTOMER_ROUTES/>}/>
-              <Route path="my-profile/*" element={<EMPLOYEE_MYPROFILE_PAGE/>}/>
-          </Routes>
-        </Box>
-    </Box>
     )
 }
